- added configure files from Prof. Ripley - added documentations for inameSDF, sdfImportSQLite, sdfImportDBI, sdflm. - added HACKING notes - added list of functions known to work in SQLiteDF.package.Rd - changed sqlite_vector accumulator to long double, fixed the accuracy problem
2006-08-18  Douglas Bates  <bates@R-project.org>
       * R/All.R (with.sqlite.data.frame, summary.sqlite.vector,
       as.logical.sqlite.vector, as.integer.sqlite.vector,
       as.double.sqlite.vector): Modify calling sequence of S3 methods to
       correspond to generic.
       * man/sqlite.data.frame.Rd: add keywords



git-svn-id: https://svn.r-project.org/R-packages/trunk@3432 694ef91d-65df-0310-b7bb-92e67a308ead

+++ SQLiteDF/src/sqlite_vector.c
@@ -427,7 +427,7 @@ SEXP sdf_variable_summary(SEXP svec, SEXP maxsum) {
 /* the global accumulator should be safe if we only do 1 cummulative or
  * aggregate at any time. it won't work for stuffs like "select max(col)-min(col)
  * from sdf_data". */
-static double g_accumulator = 0.0; /* accumulator var for cumsum, cumprod, etc. */
+static long double g_accumulator = 0.0; /* accumulator var for cumsum, cumprod, etc. */
 static int g_start = 0;            /* flag for start of cummulation */
 static int g_narm = 0;             /* for Summary group */
